Gestion delete

+++ controllers/stuff.js
@@ -1,30 +1,67 @@
 const Thing = require('../models/thing'); //Import du modèle de schéma thing 
-
+const fs = require('fs');
 
 //Fonction POST - Met tous
 exports.createThing = (req, res, next) => {
-    delete req.body._id; //Faux id envoyé par le frontend
-    const thing = new Thing({ //New instance
-        ...req.body
+    const thingObject = JSON.parse(req.body.thing);
+    delete thingObject._id;
+    delete thingObject._userId;
+    const thing = new Thing({
+        ...thingObject,
+        userId: req.auth.userId,
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}` //Recomposition de l'URL
+        //Nous utilisons req.protocol pour obtenir le premier segment (dans notre cas 'http').
+        //Nous ajoutons '://', puis utilisons req.get('host') pour résoudre l'hôte du serveur (ici, 'localhost:3000').
+        //Nous ajoutons finalement '/images/' et le nom de fichier pour compléter notre URL
     });
-    thing.save() //Enregistre les données
-        .then(() => res.status(201).json({ message: 'Objet enregistré !' })) //I l faut dire que tout va bien au cas où
-        .catch(error => res.status(400).json({ error }));
-}
+
+    thing.save()
+        .then(() => { res.status(201).json({ message: 'Objet enregistré !' }) })
+        .catch(error => { res.status(400).json({ error }) })
+};
 
 //Fonction PUT - Mettre à jour
 exports.modifyThing = (req, res, next) => {
-    Thing.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Objet modifié !' }))
-        .catch(error => res.status(400).json({ error }));
+    const thingObject = req.file ? {
+        ...JSON.parse(req.body.thing),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ...req.body };
+
+    delete thingObject._userId;
+    Thing.findOne({ _id: req.params.id })
+        .then((thing) => {
+            if (thing.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Not authorized' });
+            } else {
+                Thing.updateOne({ _id: req.params.id }, { ...thingObject, _id: req.params.id })
+                    .then(() => res.status(200).json({ message: 'Objet modifié!' }))
+                    .catch(error => res.status(401).json({ error }));
+            }
+        })
+        .catch((error) => {
+            res.status(400).json({ error });
+        });
 };
 
 //Fonction Delete - Supprimer en fonction de l'id
 exports.deleteThing = (req, res, next) => {
-    Thing.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Objet supprimé !' }))
-        .catch(error => res.status(400).json({ error }));
-}
+   Thing.findOne({ _id: req.params.id})
+       .then(thing => {
+           if (thing.userId != req.auth.userId) {
+               res.status(401).json({message: 'Not authorized'});
+           } else {
+               const filename = thing.imageUrl.split('/images/')[1];
+               fs.unlink(`images/${filename}`, () => {
+                   Thing.deleteOne({_id: req.params.id})
+                       .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                       .catch(error => res.status(401).json({ error }));
+               });
+           }
+       })
+       .catch( error => {
+           res.status(500).json({ error });
+       });
+};
 
 //Fonction GET - Récupère en fonction de l'id
 exports.getOneThing = (req, res, next) => {
